Add syncMode option to container publish

+++ scripts/Compile-App.ps1
@@ -30,7 +30,10 @@ Param(
     [switch] $skipVerification,
 
     [Parameter(Mandatory = $false)]
-    [bool] $changesOnly = $false
+    [bool] $changesOnly = $false,
+
+    [Parameter(Mandatory = $false)]
+    [string] $SyncAppMode = "Add"
 )
 
 if (-not ($credential)) {
@@ -68,7 +71,11 @@ Sort-AppFoldersByDependencies -appFolders $appFolders.Split(',') -baseFolder $bu
                 $credential = New-Object PSCredential -ArgumentList $ENV:USERNAME, $SecurePassword
             }
             Write-Host "Publishing $_"
-            Publish-BCContainerApp -containerName $containerName -appFile $appFile -skipVerification:$skipVerification -scope Tenant -sync -install -upgrade -useDevEndpoint -credential $credential
+            if ($SyncAppMode -eq "ForceSync") {
+                Publish-BCContainerApp -containerName $containerName -appFile $appFile -skipVerification:$skipVerification -scope Tenant -sync -install -upgrade -useDevEndpoint -credential $credential -syncMode ForceSync
+            } else {
+                Publish-BCContainerApp -containerName $containerName -appFile $appFile -skipVerification:$skipVerification -scope Tenant -sync -install -upgrade -useDevEndpoint -credential $credential
+            }
         } 
     }
 }
